# ICP: Improve AES-GCM performance