… module/icp/asm-x86_64/modes/THIRDPARTYLICENSE.cryptogams.descrip
… module/icp/asm-x86_64/modes/THIRDPARTYLICENSE.openssl.descrip
… module/icp/asm-x86_64/modes/aesni-gcm-x86_64.S
… module/icp/asm-x86_64/modes/ghash-x86_64.S
… tests/zfs-tests/tests/functional/rsend/send_encrypted_props.ksh