@AttilaFueloep 2019-12-19
Some further remarks:

- Since this is crypto I tried to keep my changes to the existing code at a bare minimum. This should also make it easier to merge possible future upstream changes.

- I left in  a couple of fixmes in places where I would appreciate reviewer feedback.

- The changes done to the openssl assembler routines during porting can be found [here (aesni-gcm-x86_64.S.diff)](https://gist.github.com/AttilaFueloep/c1aa86170adfbf11fdb37a2ffcad8aa3) and [here (ghash-x86_64.S.diff)](https://gist.github.com/AttilaFueloep/278697313a980f64260965c0276c40ad)

- The ICP module parameters doesn't seem to be documented, so I have no place to add documentation to.

- My testing resources are quite limited so I would really appreciate any further testing.