@behlendorf 2019-12-19
Very impressive!  This is a huge performance improvement, thank you for tackling this work.  

cc: @jasonbking @Fabian-Gruenbichler