@AttilaFueloep 2019-12-19
I'm really sorry for the build failures, but I can't reproduce them locally. Do I need to add anything besides `--enable-debug --enable-debuginfo` to configure?