@behlendorf 2019-12-19
No problem, that's why we have to CI to test across a wide range of kernel, compiler, and library versions.  It looks like you've managed to sort it out.  Using `--enable-debug --enable-debuginfo` should be sufficient.