@AttilaFueloep 2019-12-19
Yes, forgot to run `make cstyle` after my last change and my editor converted four to two spaces. `make lint` gives me tons of errors even in files I did not touch, seems Arch Linux `cppcheck` is special. But I think I can fix the errors.