@behlendorf 2019-12-19
> make lint gives me tons of errors even in files I did not touch

Try rebasing the PR on the master branch from yesterday.  A collection of cppcheck patches were merged yesterday for issues only detected by newer versions of cppcheck, up to cppcheck v1.88 should now be clean.