@AttilaFueloep 2019-12-19
Well the PR is based on master@a3640486fffc which is from today but Arch has `cppcheck 1.89-1`. I'll try downgrading it to 1.88 and see if that helps.