@behlendorf 2019-12-19
The CI is still using cppcheck 1.82 and is reporting [these](http://build.zfsonlinux.org/builders/Ubuntu%2017.10%20x86_64%20%28STYLE%29/builds/21526/steps/shell_3/logs/stdio) warnings which do look related to this change.