@AttilaFueloep 2019-12-19
Right, although I didn't fully understand the error, I tried to fix them with commit f488b0e which didn't help. I now downgraded  cppcheck to 1.86 and it's silent with and without f488b0e. May that be a false positive? Paxcheck complained about an executable stack in the assembler files, which I've already fixed but not pushed yet. Should I revert f488b0e and add a /* LINTED */ comment?