@AttilaFueloep 2019-12-19
Well, I see ztest failing in the new code. I've to have a look tomorrow.