@behlendorf 2019-12-19
That's strange, I wasn't able to reproduce the issue with cppcheck 1.88 either.  Let's revert f488b0e from the PR for now, we'll be updating the version of cppcheck used pretty shortly so this may not be an issue.