@AttilaFueloep 2019-12-19
Ok, will do tomorrow. Looks like there is no support for the `MOVBE` instruction in the testers. I'm a bit puzzled here. What virtual CPU do they use?