@behlendorf 2019-12-19
Actually, it sounds like this went pretty well since we catch this MOVBE issue.  What you're going to want to do is a check for `X86_FEATURE_MOVBE` in `include/os/linux/kernel/linux/simd_x86.h`.  This way we'll be able to detect if it's available at run time, take a look at the `zfs_*_available()` functions.  Are there other specific instructions we should be concerned about?

As for how to test this I don't think it'll be long before switching to m5 instances which according to the documentation use skylake.