@AttilaFueloep 2019-12-19
Luckily, yes. I've to go over the openssl requirements again, obviously I missed something there.

This also poses the question if this code should always be compiled in, regardless of the build environment detected by `configure` and only called on supported platforms at run time. The pclmulqdq code gets compiled in only if building on supported hardware (`defined(HAVE_PCLMULQDQ)`), I just mirrored that behaviour.