@behlendorf 2019-12-20
> should always be compiled in, regardless of the build environment detected by configure and only called on supported platforms at run time.

Yup.  And it looks like that is the case for pclmulqdq  code so your code is pretty close.  There are a couple level of checks going on here which can be a bit confusing.  But to summarize with an example, the `HAVE_PCLMULQDQ` check detects if the host build tool chain understands instruction, not if the host hardware supports it.  This is coupled with `zfs_pclmulqdq_available()` which is used for the run time check.

You'll want to add similar checks for `MOVBE` to `config/toolchain-simd.m4` and `include/os/linux/kernel/linux/simd_x86.h` and `lib/libspl/include/sys/simd.h`.  Then always compile the code in when `HAVE_MOVBE` but only enable it on module load when `zfs_pclmulqdq_available()` is true.