@AttilaFueloep 2019-12-20
@behlendorf All right, got it. Checking for tool chain support definitely makes more sense. I'll try to find some time tomorrow.