@AttilaFueloep 2019-12-22
The "signed integer overflow" lint error is a false positive and will go away with cppcheck >= 0.86. Since the testers do not run the new code now, (they are missing the `MOVBE` instruction), I don't think the test failure is related to my changes. All other tests pass, but they are not very meaningful until the testers will be updated to newer ec2 instances which will be able to test the added code.