@behlendorf 2019-12-23
@AttilaFueloep thanks for addressing the review feedback so quickly.  I agree, the testing failures you hit were unrelated to this change.

The CI has been updated to use m5d.large instances which should support the needed instruction.  I've gone ahead and resubmitted this PR for a fresh test run, but you may want to rebase it as well to clean out any old results.  Additionally, the cppcheck version has been updated that issue should be resolved.