@AttilaFueloep 2019-12-23
@behlendorf 

> The CI has been updated to use m5d.large instance

Fantastic!

Squashed and rebased, it should be ready for review now.